Stream: flatMap and fold methods need to be tested
Monad: stream monad now uses our stream implementation, not Collection's

Monoid: WC monoid copies book's answer.  Left to figure out is the count
method, and what will be in the list given to fold map with the WC
monoid.  Will it be a list of Stubs?  A list of one Part?

State: choose double in range. choose int's scale value now a double.

Gen: Prop trait/case class will be redone with chapter's given 'run' and
'forAll' methods.  Figured out Gen's union and weighted methods.  The
use of flatMap in those is hard to wrap my head around.  Specifically,
union's Gen[Int] is part of the "pipeline" of the method, but has an
outward appearance of typical, side-effectful code.

  def union[A](g1: Gen[A], g2: Gen[A]): Gen[A] = {
    val genInt: Gen[Int] = choose(0, 2)
    val genChosen: Gen[A] = genInt.flatMap {
      (i: Int) => if(i==0) g1 else g2
    }
    genChosen
  }

genInt is *not* side effecting, though.  choose returns a value
determined by the RNG passed to union's output Gen.
This is probably the pattern refered to in the book of writing steps
to be interpreted later:

"""
   We saw in the case of buyCoffee how we could separate the creation of the Charge from the interpretation or processing of that Charge. In general, we’ll learn how this sort of transformation can be applied to any function with side effects to push these effects to the outer layers of the program. Functional programmers often speak of implementing programs with a pure core and a thin layer on the outside that handles effects.
"""

Gen/Prop relies on State, Par and Stream.
Monoid relies on Par and Gen/Prop.